+++ app/src/main/java/com/mobile/xplore_nu/XploreNuApplication.kt
@@ -1,0 +1,5 @@
+package com.mobile.xplore_nu
+
+import android.app.Application
+
+class XploreNuApplication : Application()
