--- app/src/main/java/com/nbapps/xplore_nu/XploreNuApplication.kt
@@ -1,6 +1,0 @@
-package com.nbapps.xplore_nu
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-class XploreNuApplication : Application()
