+++ data/src/main/java/com/mobile/data/local/models/DUser.kt
@@ -1,0 +1,7 @@
+package com.mobile.data.local.models
+
+data class DUser(
+    val id: String,
+    val name: String,
+    val email: String
+)
