--- data/src/main/java/com/nbapps/data/local/models/DUser.kt
@@ -1,7 +1,0 @@
-package com.nbapps.data.local.models
-
-data class DUser(
-    val id: String,
-    val name: String,
-    val email: String
-)
