+++ domain/build.gradle.kts
@@ -8,6 +8,10 @@ java {
     sourceCompatibility = JavaVersion.VERSION_17
     targetCompatibility = JavaVersion.VERSION_17
 }
+tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile>().configureEach {
+    kotlinOptions.jvmTarget = "17"
+}
+
 
 dependencies {
     implementation(libs.kotlinx.coroutines.core)
