+++ domain/src/main/java/com/mobile/domain/models/User.kt
@@ -1,0 +1,7 @@
+package com.mobile.domain.models
+
+data class User(
+    val id: String,
+    val name: String,
+    val email: String
+)
