--- domain/src/main/java/com/nbapps/domain/models/User.kt
@@ -1,7 +1,0 @@
-package com.nbapps.domain.models
-
-data class User(
-    val id: String,
-    val name: String,
-    val email: String
-)
